Add docker make commands and CI

+++ cmd/unused-exporter/exporter.go
@@ -182,7 +182,7 @@ func (e *exporter) pollProvider(p unused.Provider) {
 			}
 
 			addMetric(e.info, 1)
-			addMetric(e.dur, int64(dur.Microseconds()))
+			addMetric(e.dur, dur.Microseconds())
 			addMetric(e.suc, success)
 
 			for ns, di := range diskInfoByNamespace {
